Introducing new setting 'maxPageSize' to enhance data retrieval capability.
I missed a lot of scripting functions when loading big business services. So this commit introduces a new setting called 'maxPageSize' which allows users to specify the maximum number of records to be loaded per API call. By default, the Siebel API loads only 10 records, but with this new setting, users can adjust this limit according to their requirements.